Prevent recursion when filling a variable with $copymerge and $join
Otherwise this leads to a repeated values on refresh (when the scripts
are run again) because $copymerge does not see the already joined
values as a duplicate and hence can not deduplicate it.
Also do not rejoin an already joined value (which starts with a space).

Example history of the variable:
1) "A; B" after the initial population with $copymerge
2) " [A B]" after the initial $join from `Filename Components.pts`
3a) " [ [A B]]" when running `Filename Components.pts` again, or...
3b) " [A B]; A; B" after the repeated population
4b) "[ [A B] A B]" after the repeated join ...