Do not replace Unicode punctuation in artist tags
There are just too many different tags, and the solution would not work
for the performer:* tags anyway.
Changing only the tags which are used for file naming would lead to a
library with duplicate artists.

So instead of doing these replacements for artist tags, we will only do
them for title tags and in the file naming script.